[winsvc] Restructure and introduce fake test pkg
This commit breaks the winsvc pkg into separate components for Windows
services (winsvc) and the overlying service manager interface (manager),
as well as a separate fake package to test both.
